Move function_pointers.h to protocol_layer
