+++ standalone/include/psen_scan_v2_standalone/protocol_layer/function_pointers.h
@@ -23,11 +23,14 @@
 
 namespace psen_scan_v2_standalone
 {
+namespace protocol_layer
+{
 using SendRequestCallback = std::function<void()>;
 using ReplyCallback = std::function<void()>;
 using MonitoringFrameCallback = std::function<void(const data_conversion_layer::monitoring_frame::Message&)>;
 using ErrorCallback = std::function<void(const std::string&)>;
 using LaserScanCallback = std::function<void(const api::LaserScan&)>;
+}  // namespace protocol_layer
 }  // namespace psen_scan_v2_standalone
 
 #endif  // PSEN_SCAN_V2_STANDALONE_FUNCTION_POINTERS_H
